Delete fix

+++ src/main/java/ch/supsi/webapp/web/model/BlogPostRepository.java
@@ -3,6 +3,6 @@ package ch.supsi.webapp.web.model;
 import org.springframework.data.repository.CrudRepository;
 
 public interface BlogPostRepository extends CrudRepository<BlogPost, Integer> {
-    BlogPost findTop1ById(int id);
+    BlogPost findTop1ById(long id);
     BlogPost findFirstByOrderByIdDesc();
 }
